requirements.txt
requests
ttkthemes
Pillow
tkintermapview
python-dotenv
pandas
matplotlib
psycopg2-binary
Flask-WTF
